Changes from ICRA2023 submission (#168)
* Update viewset to contain 50 viewpoints

* Fix termination criteria

* Fix airsim client

* Fix coverage planner

* Update grid planner

This PR updates the grid planner

* Record image count in benchmark results

* Add viewpoint noise

* Resolve inheritance problems with spherical coverage map

* Add exploration map

Add view utility map

* Configure offset polygon for coverage planning

* Enable greedy for airsim client

* Consider altitude offset when turning

* Add colored mesh if mesh texture is available

* Random viewpoint generation


Fix

* Log raw benchmark outputs

* Write raw vehicle path

* Fix attitude outputs

* Coverage cleanup

* Output relevant information for visualization

+++ adaptive_viewutility/include/adaptive_viewutility/common.h
@@ -163,7 +163,8 @@ bool parseAttitudeFromText(std::string text_path, std::string image_file, Eigen:
 }
 
 double getRandom(double min, double max) {
-  return std::abs(max - min) * static_cast<double>(rand()) / static_cast<double>(RAND_MAX) + std::min(max, min);
+  return std::abs(max - min) * static_cast<double>(static_cast<double>(rand()) / static_cast<double>(RAND_MAX)) +
+         std::min(max, min);
 }
 
 double getNormalRandom(double mean, double sigma) {
